Check untrusted user-ids
Based on a careful analysis of when user-ids are verified, we've identified
some places where untrusted user-ids were treated as trusted and fixed those.

Add an IsAdmin field so that during the GetElections call the UI can find out
if the user is an admin or not.
